fix double to cgfloat issue

+++ ios/Classes/Annotations/AnnotationController.swift
@@ -144,7 +144,7 @@ class AnnotationController: NSObject {
                 })
             }
             if let zPosition = newAnnotation.zPosition, oldAnnotation.zPosition != zPosition {
-                oldAnnotationView.layer.zPosition = zPosition
+                oldAnnotationView.layer.zPosition = CGFloat(zPosition)
             }
             oldAnnotation.coordinate = newAnnotation.coordinate
             oldAnnotation.icon = newAnnotation.icon
